requirements.txt
simplejson>=3.17.3
lxml>=4.6.3
codecov
coverage
flask>=2.0.2
pre-commit
pytest>=6.2.4
PyJWT>=2.1.0
requests>=2.26.0
pymysql>=1.0.2
click>=8.0.3
werkzeug>=2.0.2
setuptools>=58.0.4
DBUtils>=2.0
cryptography
matplotlib